Reduce memory footprint from indexed dictionaries
Dictionaries are indexed using a map with the hex-string representation
of the hashed password as key, and the password itself as value.

The hex-string representation of a byte requires two hex digits which
takes two bytes to store. The space needed to store these keys in memory
can be halved by not storing it hex-encoded.

Use the bytes (as string) for the key to prevent wasting memory.

This is especially important when using very large dictionaries and/or
hash functions that return relatively big hashes, e.g. sha512.